# swagger.json
{
   "swagger": "2.0",
   "info": {
      "title": "Hotel API",
      "description": "API for hotel management",
      "version": "1.0"
   },
   "produces": [
      "application/json"
   ],
   "host": "127.0.0.1:3000",
   "basePath": "/",
   "paths": {
...
      "/services": {
         "get": {
            "operationId": "services",
            "tags": [
               "Services"
            ],
            "description": "Récupère la liste des services",
            "parameters": [],
            "responses": {
               "200": {
                  "description": "successful operation",
                  "schema": {
                     "type": "array",
                     "items": {
                        "type": "object",
                        "properties": {
                           "id_service": {
                              "type": "integer"
                           },
                           "service_name": {
                              "type": "string"
                           },
                           "description": {
                              "type": "string"
                           },
                           "price": {
                              "type": "number",
                              "format": "float"
                           }
                        }
                     }
                  }
               },
               "404": {
                  "description": "Not found"
               },
               "500": {
                  "description": "Internal server error"
               }
            }
         },
         "post" : {
            "operationId": "services_create",
            "tags": [
               "Services"
            ],
            "description": "Crée un service",
            "parameters": [
               {
                  "name": "service",
                  "in": "body",
                  "required": true,
                  "schema": {
                     "type": "object",
                     "properties": {
                        "service_name": {
                           "type": "string"
                        },
                        "description": {
                           "type": "string"
                        },
                        "price": {
                           "type": "number",
                           "format": "float"
                        }
                     }
                  }
               }
            ],
            "responses": {
               "201": {
                  "description": "successful operation",
                  "schema": {
                     "type": "object",
                     "properties": {
                        "id_service": {
                           "type": "integer"
                        },
                        "service_name": {
                           "type": "string"
                        },
                        "description": {
                           "type": "string"
                        },
                        "price": {
                           "type": "number",
                           "format": "float"
                        }
                     }
                  }
               }
            }
         }
      },
...
      "/payments": {
         "get": {
            "operationId": "payments",
            "tags": [
               "Payments"
            ],
            "description": "Récupère la liste des paiements",
            "parameters": [],
            "responses": {
               "200": {
                  "description": "successful operation",
                  "schema": {
                     "type": "array",
                     "items": {
                        "type": "object",
                        "properties": {
                           "id_payment": {
                              "type": "integer"
                           },
                           "payment_date": {
                              "type": "string",
                              "format": "date-time"
                           },
                           "amount": {
                              "type": "number",
                              "format": "float"
                           },
                           "payment_method": {
                              "type": "string"
                           }
                        }
                     }
                  }
               },
               "404": {
                  "description": "Not found"
               },
               "500": {
                  "description": "Internal server error"
               }
            }
         },
         "post" : {
            "operationId": "payments_create",
            "tags": [
               "Payments"
            ],
            "description": "Crée un paiement",
            "parameters": [
               {
                  "name": "payment",
                  "in": "body",
                  "required": true,
                  "schema": {
                     "type": "object",
                     "properties": {
                        "payment_date": {
                           "type": "string",
                           "format": "date-time"
                        },
                        "amount": {
                           "type": "number",
                           "format": "float"
                        },
                        "payment_method": {
                           "type": "string"
                        }
                     }
                  }
               }
            ],
            "responses": {
               "201": {
                  "description": "successful operation",
                  "schema": {
                     "type": "object",
                     "properties": {
                        "id_payment": {
                           "type": "integer"
                        },
                        "payment_date": {
                           "type": "string",
                           "format": "date-time"
                        },
                        "amount": {
                           "type": "number",
                           "format": "float"
                        },
                        "payment_method": {
                           "type": "string"
                        }
                     }
                  }
               }
            }
         }
      },
...
      "/inclures": {
         "get": {
            "operationId": "inclures",
            "tags": [
               "Inclures"
            ],
            "description": "Récupère la liste des inclures",
            "parameters": [],
            "responses": {
               "200": {
                  "description": "successful operation",
                  "schema": {
                     "type": "array",
                     "items": {
                        "type": "object",
                        "properties": {
                           "id_inclure": {
                              "type": "integer"
                           },
                           "id_service": {
                              "type": "integer"
                           },
                           "id_reservation": {
                              "type": "integer"
                           },
                           "quantity": {
                              "type": "integer"
                           },
                           "total_price": {
                              "type": "number",
                              "format": "float"
                           }
                        }
                     }
                  }
               },
               "404": {
                  "description": "Not found"
               },
               "500": {
                  "description": "Internal server error"
               }
            }
         },
         "post": {
            "operationId": "inclures_create",
            "tags": [
               "Inclures"
            ],
            "description": "Crée un inclure",
            "parameters": [
               {
                  "name": "inclure",
                  "in": "body",
                  "required": true,
                  "schema": {
                     "type": "object",
                     "properties": {
                        "id_service": {
                           "type": "integer"
                        },
                        "id_reservation": {
                           "type": "integer"
                        },
                        "quantity": {
                           "type": "integer"
                        },
                        "total_price": {
                           "type": "number",
                           "format": "float"
                        }
                     }
                  }
               }
            ],
            "responses": {
               "201": {
                  "description": "successful operation",
                  "schema": {
                     "type": "object",
                     "properties": {
                        "id_inclure": {
                           "type": "integer"
                        },
                        "id_service": {
                           "type": "integer"
                        },
                        "id_reservation": {
                           "type": "integer"
                        },
                        "quantity": {
                           "type": "integer"
                        },
                        "total_price": {
                           "type": "number",
                           "format": "float"
                        }
                     }
                  }
               }
            }
         }
      },
...
   }
}